@@ -18,10 +18,10 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "twig/twig": "~1.30",
-        "illuminate/support": "5.0.*|5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*",
-        "illuminate/view": "5.0.*|5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*"
+        "php": ">=7.1",
+        "twig/twig": "~2.0",
+        "illuminate/support": "5.5.*|5.6.*|5.7.*",
+        "illuminate/view": "5.5.*|5.6.*|5.7.*"
     },
     "require-dev": {
         "laravel/framework": "5.0.*",
@@ -42,7 +42,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "0.10-dev"
+            "dev-master": "0.11-dev"
         },
         "laravel": {
             "providers": [
