@@ -2,15 +2,15 @@
 
 namespace TwigBridge\Extension\Bridge;
 
-use Twig_Extension;
-use Twig_SimpleFunction;
+use Twig\TwigFunction;
+use Twig\Extension\AbstractExtension;
 
 /**
  * Use Former in your Twig templates.
  *
  * @see https://github.com/formers/former
  */
-class Former extends Twig_Extension
+class Former extends AbstractExtension
 {
 	/**
 	 * @var \Former\Former
@@ -41,7 +41,7 @@ class Former extends Twig_Extension
 	public function getFunctions()
 	{
 		return [
-		  new Twig_SimpleFunction(
+		  new TwigFunction(
 			'former_*',
 			function ($name) {
 				$arguments = array_slice(func_get_args(), 1);
