@@ -2,8 +2,12 @@
 
 namespace TwigBridge\Extension\Laravel;
 
-use Illuminate\Support\Debug\HtmlDumper;
+use Symfony\Component\VarDumper\Dumper\HtmlDumper;
 use Symfony\Component\VarDumper\Cloner\VarCloner;
+use Twig\Environment;
+use Twig\Template;
+use Twig\TwigFunction;
+use Twig\Extension\AbstractExtension;
 
 /**
  * Dump a variable or the view context
@@ -12,7 +16,7 @@ use Symfony\Component\VarDumper\Cloner\VarCloner;
  * @see https://github.com/symfony/symfony/blob/2.6/src/Symfony/Bridge/Twig/Extension/DumpExtension.php
  * @author Nicolas Grekas <p@tchwork.com>
  */
-class Dump extends \Twig_Extension
+class Dump extends AbstractExtension
 {
     public function __construct()
     {
@@ -22,7 +26,7 @@ class Dump extends \Twig_Extension
     public function getFunctions()
     {
         return array(
-          new \Twig_SimpleFunction('dump', array($this, 'dump'), array('is_safe' => array('html'), 'needs_context' => true, 'needs_environment' => true)),
+          new TwigFunction('dump', array($this, 'dump'), array('is_safe' => array('html'), 'needs_context' => true, 'needs_environment' => true)),
         );
     }
 
@@ -31,7 +35,7 @@ class Dump extends \Twig_Extension
         return 'TwigBridge_Extension_Laravel_Dump';
     }
 
-    public function dump(\Twig_Environment $env, $context)
+    public function dump(Environment $env, $context)
     {
         if (!$env->isDebug()) {
             return;
@@ -39,7 +43,7 @@ class Dump extends \Twig_Extension
         if (2 === func_num_args()) {
             $vars = array();
             foreach ($context as $key => $value) {
-                if (!$value instanceof \Twig_Template) {
+                if (!$value instanceof Template) {
                     $vars[$key] = $value;
                 }
             }
