@@ -2,10 +2,10 @@
 
 namespace TwigBridge\Extension\Laravel\Legacy;
 
-use Twig_Extension;
-use Twig_Function_Function;
 use Illuminate\Config\Repository as Config;
 use Illuminate\Foundation\Application;
+use Twig\TwigFunction;
+use Twig\Extension\AbstractExtension;
 
 /**
  * Handles undefined function calls in a Twig template by checking
@@ -14,7 +14,7 @@ use Illuminate\Foundation\Application;
  *
  * {{ auth_check() }} would call Auth::check()
  */
-class Facades extends Twig_Extension
+class Facades extends AbstractExtension
 {
     /**
      * @var \Illuminate\Foundation\Application
@@ -175,7 +175,7 @@ class Facades extends Twig_Extension
      *
      * @param string $name Function name.
      *
-     * @return Twig_Function_Function|false
+     * @return TwigFunction|false
      */
     public function getLookup($name)
     {
@@ -187,12 +187,12 @@ class Facades extends Twig_Extension
     /**
      * Add undefined function to the cache.
      *
-     * @param string                 $name     Function name.
-     * @param Twig_Function_Function $function Function to cache.
+     * @param string         $name     Function name.
+     * @param TwigFunction $function Function to cache.
      *
      * @return void
      */
-    public function setLookup($name, Twig_Function_Function $function)
+    public function setLookup($name, TwigFunction $function)
     {
         $this->lookup[strtolower($name)] = $function;
     }
@@ -202,7 +202,7 @@ class Facades extends Twig_Extension
      *
      * @param string $name Undefined function name.
      *
-     * @return Twig_Function_Function|false
+     * @return TwigFunction|false
      */
     public function getFunction($name)
     {
@@ -227,7 +227,7 @@ class Facades extends Twig_Extension
         if (array_key_exists($class, $this->aliases)) {
 
             $class    = $this->aliases[$class];
-            $function = new Twig_Function_Function($class.'::'.$method);
+            $function = new TwigFunction($name, $class.'::'.$method);
 
             $this->setLookup($name, $function);
 
