@@ -2,13 +2,13 @@
 
 namespace TwigBridge\Twig;
 
-use Twig_Template;
 use Illuminate\View\View;
+use Twig\Template as TwigTemplate;
 
 /**
  * Default base class for compiled templates.
  */
-abstract class Template extends Twig_Template
+abstract class Template extends TwigTemplate
 {
     /**
      * @var bool Have the creator/composer events fired.
@@ -90,34 +90,5 @@ abstract class Template extends Twig_Template
     public function setFiredEvents($fired = true)
     {
         $this->firedEvents = $fired;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function getAttribute(
-        $object,
-        $item,
-        array $arguments = [],
-        $type = Twig_Template::ANY_CALL,
-        $isDefinedTest = false,
-        $ignoreStrictCheck = false
-    ) {
-        // We need to handle accessing attributes on an Eloquent instance differently
-        if (Twig_Template::METHOD_CALL !== $type and is_a($object, 'Illuminate\Database\Eloquent\Model')) {
-            // We can't easily find out if an attribute actually exists, so return true
-            if ($isDefinedTest) {
-                return true;
-            }
-
-            if ($this->env->hasExtension('sandbox')) {
-                $this->env->getExtension('sandbox')->checkPropertyAllowed($object, $item);
-            }
-
-            // Call the attribute, the Model object does the rest of the magic
-            return $object->$item;
-        } else {
-            return parent::getAttribute($object, $item, $arguments, $type, $isDefinedTest, $ignoreStrictCheck);
-        }
     }
 }
