@@ -23,7 +23,7 @@ class BridgeTest extends Base
     public function testSetLexer()
     {
         $app = $this->getApplication();
-        $app['twig.lexer'] = m::mock('Twig_LexerInterface');
+        $app['twig.lexer'] = m::mock('Twig_Lexer');
         $app['twig.lexer']->shouldReceive('fooBar')->andReturn('baz');
         $app['twig.extensions'] = [];
 
@@ -31,7 +31,7 @@ class BridgeTest extends Base
         $provider->register();
         $provider->boot();
 
-        $this->assertEquals($app['twig']->getLexer()->fooBar(), 'baz');
+        $this->assertEquals($this->readAttribute($app['twig'], 'lexer')->fooBar(), 'baz');
     }
 
     public function testAddExtensions()
